+++ db/migrate/20210601042208_add_batch_ingest_items.rb
@@ -1,0 +1,5 @@
+class AddBatchIngestItems < ActiveRecord::Migration[6.0]
+  def change
+    add_reference :items, :batch_ingest, foreign_key: true
+  end
+end